About the Package

This package allows a user to generate noisy correlation functions from real axis distributions. The noise is correlated using the technique developed by Hui Shao et al PhysRevX.7.041072

API

GenerateCorrelationFunctions

AppendDistribution!

SynthAC.GenerateCorrelationFunctions — Function

GenerateCorrelationFunctions(DistributionArray,β,Δτ,fermionic;
                             outfile="",NBins=100,AutoCorrelationTime=0.5,
                             σ0=0.005,Maxω=10.0)

Generates synthetic correlation functions from the passed in functions in the distribution array. This generates bins of data with correlated errors using the method developed by Hui Shao, Yan Qi Qin, Sylvain Capponi, Stefano Chesi, Zi Yang Meng, and Anders W. Sandvik (PHYSICAL REVIEW X 7, 041072 (2017))

Arguments

DistributionArray: Array of Anonymous Functions

β: Inverse Temperature

Δτ: Timestep in imaginary time

fermionic: Boolean, true if using fermionic kernel

Optinal Arguments

outfile: file to save dictionary to. .jld2 format is recommended. "" for no output file

NBins: Number of bins of noisy correlation functions to generate

AutoCorrelationTime: Auto correlation length in imaginary time

σ0: Tuneable parameter to increase/decrease variance in correlation functions

Maxω: range to integrate over. (-∞,∞) is not yet supported, and exponential values in kernels limit range possible

Nωn: Number of Matsubara frequency points

Blurtype: Either "gamma" or "absgauss". "gamma" uses a gamma distribution which does not go below zero. "absgauss" uses a normal distribution and takes the absolute value

Returns

Dict{String,Any}(...) containing the keys

"A": True distribution

`"ωs": Grid of ω values for A

"β": Inverse temperature

"τs": Imaginary time grid

"ωns": Matsubara frequency grid

"ξ": Autocorrelation Time

"σ0": Tuneable uncorrelated standard error parameter

"Gτ": Array of dimensions [NBins,size(τs,1)]

"Gτ_calc": Reference correlation function without noise

"Gωn": Array of dimensions [NBins,size(ωns,1)]

"Gωn_calc": Reference correlation function without noise

SynthAC.AppendDistribution! — Function

AppendDistribution!(DistributionArray,Distribution)

Appends a new anonymous function to array of anonymous functions

Arguments

DistributionArray: Array to append anonymous function to. May be empty array

Distribution: Anonymous function to apped